msmnile: Move the kernel flags to msmnile.mk
The configuration (compile with LLVM) in AndroidBoard.mk is
picked up only during kernel compilation. However, some tech
packages (such as audio and wlan) are placed out of the kernel
tree under vendor/qcom. Apparently, the configurations are not
picked up when these tech packages are being built. Hence,
moving the configuration to msmnile.mk such that the standalone
kernel compilation and the tech packages compilation would be possible.

Change-Id: Ibcbd1e443244a40eaafa69042ad4573d4dff6704