raphael: fod: support dispparam handling mechanism as fallback
On previous fastpath-try with non OSS vendor, we encounter an issues that
DisplayFeature can't track the state of it for PCC tuning, which can
leads panel became out-of-sync on DC dimming enabled.
But since DisplayFeature is gone (and probably, PCC tuning will be
handled directly via kernelspace), firing HBM via disp_param is safe to
be used now as fallback if there's no fod_hbm node.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>